+++ nallely/core/virtual_device.py
@@ -731,6 +731,22 @@ class VirtualDevice(threading.Thread):
     def nonstream_links(self):
         return self.links[int(False)]
 
+    @property
+    def outgoing_links(self):
+        return list(self.links_registry.values())
+
+    @property
+    def incoming_links(self):
+        from .world import all_devices
+
+        links = []
+        self_repr = f"{self.uuid}::"
+        for device in all_devices():
+            for (_, dst), link in device.links_registry.items():
+                if dst.startswith(self_repr):
+                    links.append(link)
+        return links
+
     def scale(self, min=None, max=None, method="lin", as_int=False):
         return Scaler(
             self,